requirements.txt
matplotlib==3.3.4
python_igraph==0.9.8
cmake==3.22.0
